Use c_int at the CUBA library boundary so ILP64 builds work
libcuba is C code compiled with 32-bit ints, but Cuba_Integrate and the
integrand callbacks used default integers. POLARSO=yes builds compile
with 8-byte default integers (-i8/-fdefault-integer-8 for ILP64 MKL),
so every CUBA-based emission mode segfaulted inside the integrand
callback in those builds - field emission v2 among them, which made
POLARSO unusable for field emission.

Every integer crossing into libcuba is now integer(c_int):
- Cuba_Integrate keeps its default-integer dummies (callers unchanged)
  and passes c_int copies of nvec/userdata/mineval/maxeval, reading
  nregions/neval/fail back through c_int locals.
- The Cuba argument parameters (ndim, ncomp, flags, seed, keys, ...)
  are c_int.
- All eleven integrand callbacks declare their integer arguments and
  results as c_int, with default-kind local copies where userdata/nvec
  are passed on to regular Fortran procedures.

In default builds c_int equals the default integer kind, so nothing
changes there.

Verified: the full test suite (155 tests, including the Cuba
Suave/Divonne/Cuhre integration tests against analytic values) passes
in both the default nvfortran build and the ILP64 ifx POLARSO build,
and the field-emission-v2 + POLARSO smoke run that previously
segfaulted in integrand_cuba_fe_v now completes cleanly.

+++ src/mod_emission_tip.f90
@@ -1769,11 +1769,14 @@ end function Elec_supply_tip
   ! ----------------------------------------------------------------------------
   ! The integration function for the Cuba library
   !
-  integer function integrand_cuba_fe_tip_test(ndim, xx, ncomp, ff, userdata)
+  integer(c_int) function integrand_cuba_fe_tip_test(ndim, xx, ncomp, ff, userdata)
+    ! Called from the C Cuba library: the integer arguments and the result
+    ! must be integer(c_int) (see mod_cuba_integration)
+    use, intrinsic :: iso_c_binding, only: c_int
     ! Input / output variables
-    integer, intent(in) :: ndim ! Number of dimensions (Should be 2)
-    integer, intent(in) :: ncomp ! Number of vector-components in the integrand (Always 1 here)
-    integer, intent(in) :: userdata ! Additional data passed to the integral function (In our case the number of the emitter)
+    integer(c_int), intent(in) :: ndim ! Number of dimensions (Should be 2)
+    integer(c_int), intent(in) :: ncomp ! Number of vector-components in the integrand (Always 1 here)
+    integer(c_int), intent(in) :: userdata ! Additional data passed to the integral function (In our case the number of the emitter)
     double precision, intent(in), dimension(1:ndim)   :: xx ! Integration points, between 0 and 1
     double precision, intent(out), dimension(1:ncomp) :: ff ! Results of the integrand function
 
@@ -1833,10 +1836,13 @@ end function Elec_supply_tip
 
   ! ----------------------------------------------------------------------------
   ! This subroutine does the Cuba integration for the GTF emission tip.
-  integer function integrand_cuba_gtf_tip(ndim, xx, ncomp, ff, userdata)
-    integer, intent(in) :: ndim ! Number of dimensions (Should be 2)
-    integer, intent(in) :: ncomp ! Number of vector-components in the integrand (Always 1 here)
-    integer, intent(in) :: userdata ! Additional data passed to the integral function (In our case the number of the emitter)
+  integer(c_int) function integrand_cuba_gtf_tip(ndim, xx, ncomp, ff, userdata)
+    ! Called from the C Cuba library: the integer arguments and the result
+    ! must be integer(c_int) (see mod_cuba_integration)
+    use, intrinsic :: iso_c_binding, only: c_int
+    integer(c_int), intent(in) :: ndim ! Number of dimensions (Should be 2)
+    integer(c_int), intent(in) :: ncomp ! Number of vector-components in the integrand (Always 1 here)
+    integer(c_int), intent(in) :: userdata ! Additional data passed to the integral function (In our case the number of the emitter)
     double precision, intent(in), dimension(1:ndim)   :: xx ! Integration points, between 0 and 1
     double precision, intent(out), dimension(1:ncomp) :: ff ! Results of the integrand function
